code/requirements.txt
Django==4.2.20
django-simpleui==2025.5.17
djangorestframework==3.16.0